app/requirements.txt
dnspython==2.0.0
dominate==2.5.1
eventlet==0.25.2
Flask==0.10.1
Flask-Assets==2.0
Flask-Bootstrap4==4.0.2
Flask-Login==0.5.0
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gevent==20.6.2
gevent-websocket==0.10.1
greenlet==0.4.16
gunicorn==20.0.4
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
monotonic==1.5
PyMySQL==0.10.0
python-engineio==3.13.1
python-socketio==4.6.0
redis==3.5.3
six==1.15.0
SQLAlchemy==1.3.18
visitor==0.1.3
webassets==2.0
Werkzeug==0.9.6
WTForms==2.3.1
zope.event==4.4
zope.interface==5.1.0